Fetch and validation of project is performed by backend directly
Move to utils fetch and validate methods
