+++ mlflow/projects/__init__.py
@@ -4,7 +4,6 @@ The ``mlflow.projects`` module provides an API for running MLflow projects local
 
 from __future__ import print_function
 
-from distutils import dir_util
 import hashlib
 import json
 import yaml
@@ -20,17 +19,13 @@ import docker
 
 from mlflow.projects.backend.loader import load_backend
 import mlflow.tracking as tracking
-import mlflow.tracking.fluent as fluent
-from mlflow.entities import RunStatus, SourceType
+from mlflow.entities import RunStatus
 from mlflow.exceptions import ExecutionException, MlflowException
-from mlflow.projects import _project_spec
 from mlflow.projects.submitted_run import LocalSubmittedRun, SubmittedRun
 from mlflow.projects.utils import (
-    _is_file_uri, _is_local_uri, _is_valid_branch_name, _is_zip_uri,
-    _get_git_repo_url, _parse_subdirectory, _expand_uri, _get_storage_dir,
-    _GIT_URI_REGEX,
+    _is_valid_branch_name, _get_git_repo_url, _expand_uri, _get_storage_dir,
+    prepare_project_and_run
 )
-from mlflow.tracking.context.default_context import _get_user
 from mlflow.tracking.context.git_context import _get_git_commit
 from mlflow.tracking.fluent import _get_experiment_id
 from mlflow.store.artifact.artifact_repository_registry import get_artifact_repository
@@ -42,10 +37,9 @@ from mlflow.store.artifact.s3_artifact_repo import S3ArtifactRepository
 from mlflow.utils import databricks_utils, file_utils, process
 from mlflow.utils.file_utils import path_to_local_sqlite_uri, path_to_local_file_uri
 from mlflow.utils.mlflow_tags import (
-    MLFLOW_PROJECT_ENV, MLFLOW_DOCKER_IMAGE_URI, MLFLOW_DOCKER_IMAGE_ID, MLFLOW_USER,
-    MLFLOW_SOURCE_NAME, MLFLOW_SOURCE_TYPE, MLFLOW_GIT_COMMIT, MLFLOW_GIT_REPO_URL,
-    MLFLOW_GIT_BRANCH, LEGACY_MLFLOW_GIT_REPO_URL, LEGACY_MLFLOW_GIT_BRANCH_NAME,
-    MLFLOW_PROJECT_ENTRY_POINT, MLFLOW_PARENT_RUN_ID, MLFLOW_PROJECT_BACKEND,
+    MLFLOW_PROJECT_ENV, MLFLOW_DOCKER_IMAGE_URI, MLFLOW_DOCKER_IMAGE_ID,
+    MLFLOW_GIT_REPO_URL, MLFLOW_GIT_BRANCH, LEGACY_MLFLOW_GIT_REPO_URL,
+    LEGACY_MLFLOW_GIT_BRANCH_NAME, MLFLOW_PROJECT_BACKEND,
 )
 from mlflow.utils.uri import get_db_profile_from_uri, is_databricks_uri
 from mlflow.projects import backend, utils
@@ -100,109 +94,85 @@ def _run(uri, experiment_id, entry_point="main", version=None, parameters=None,
     Helper that delegates to the project-running method corresponding to the passed-in backend.
     Returns a ``SubmittedRun`` corresponding to the project run.
     """
-
-    parameters = parameters or {}
-    work_dir = _fetch_project(uri=uri, force_tempdir=False, version=version)
-    project = _project_spec.load_project(work_dir)
-    _validate_execution_environment(project, backend_name)
-    project.get_entry_point(entry_point)._validate_parameters(parameters)
-
     if not backend_config:
         backend_config = dict()
 
     backend = load_backend(backend_name=backend_name)
-    if backend:
-        backend.validate_backend_config(backend_config)
-
-    if run_id:
-        active_run = tracking.MlflowClient().get_run(run_id)
-    else:
-        active_run = _create_run(uri, experiment_id, work_dir, entry_point)
-
-    # Consolidate parameters for logging.
-    # `storage_dir` is `None` since we want to log actual path not downloaded local path
-    entry_point_obj = project.get_entry_point(entry_point)
-    final_params, extra_params = entry_point_obj.compute_parameters(parameters, storage_dir=None)
-    for key, value in (list(final_params.items()) + list(extra_params.items())):
-        tracking.MlflowClient().log_param(active_run.info.run_id, key, value)
-
-    repo_url = _get_git_repo_url(work_dir)
-    if repo_url is not None:
-        for tag in [MLFLOW_GIT_REPO_URL, LEGACY_MLFLOW_GIT_REPO_URL]:
-            tracking.MlflowClient().set_tag(active_run.info.run_id, tag, repo_url)
-
-    # Add branch name tag if a branch is specified through -version
-    if _is_valid_branch_name(work_dir, version):
-        for tag in [MLFLOW_GIT_BRANCH, LEGACY_MLFLOW_GIT_BRANCH_NAME]:
-            tracking.MlflowClient().set_tag(active_run.info.run_id, tag, version)
 
     if backend:
-        run_id = active_run.info.run_id
-        tracking.MlflowClient().set_tag(run_id, MLFLOW_PROJECT_BACKEND,
-                                        backend.name)
-        return backend.run(run_id=run_id, project_uri=uri, entry_point=entry_point,
-                           params=parameters, backend_config=backend_config,
-                           project_dir=work_dir)
+        submitted_run = backend.run(run_id=run_id, experiment_id=experiment_id, project_uri=uri,
+                                    entry_point=entry_point, params=parameters, version=version,
+                                    backend_config=backend_config)
+        tracking.MlflowClient().set_tag(submitted_run.run_id,
+                                        MLFLOW_PROJECT_BACKEND, backend.name)
+        return submitted_run
 
+    fetched_project = prepare_project_and_run(
+        uri, experiment_id, run_id, entry_point, parameters, version)
     if backend_name == "local" or backend_name is None:
-        tracking.MlflowClient().set_tag(active_run.info.run_id,
+        tracking.MlflowClient().set_tag(fetched_project.run.info.run_id,
                                         MLFLOW_PROJECT_BACKEND, "local")
         command_args = []
         command_separator = " "
         # If a docker_env attribute is defined in MLproject then it takes precedence over conda yaml
         # environments, so the project will be executed inside a docker container.
-        if project.docker_env:
-            tracking.MlflowClient().set_tag(active_run.info.run_id, MLFLOW_PROJECT_ENV,
+        if fetched_project.project.docker_env:
+            tracking.MlflowClient().set_tag(fetched_project.run.info.run_id, MLFLOW_PROJECT_ENV,
                                             "docker")
-            _validate_docker_env(project)
+            _validate_docker_env(fetched_project.project)
             _validate_docker_installation()
-            image = _build_docker_image(work_dir=work_dir,
-                                        repository_uri=project.name,
-                                        base_image=project.docker_env.get('image'),
-                                        run_id=active_run.info.run_id)
-            command_args += _get_docker_command(image=image, active_run=active_run,
-                                                volumes=project.docker_env.get("volumes"),
-                                                user_env_vars=project.docker_env.get("environment"))
+            image = _build_docker_image(work_dir=fetched_project.work_dir,
+                                        repository_uri=fetched_project.project.name,
+                                        base_image=fetched_project.project.docker_env.get('image'),
+                                        run_id=fetched_project.run.info.run_id)
+            user_env_vars = fetched_project.project.docker_env.get("environment")
+            command_args += _get_docker_command(image=image, active_run=fetched_project.run,
+                                                volumes=fetched_project.project.docker_env.get(
+                                                    "volumes"),
+                                                user_env_vars=user_env_vars)
         # Synchronously create a conda environment (even though this may take some time)
         # to avoid failures due to multiple concurrent attempts to create the same conda env.
         elif use_conda:
-            tracking.MlflowClient().set_tag(active_run.info.run_id, MLFLOW_PROJECT_ENV, "conda")
+            tracking.MlflowClient().set_tag(fetched_project.run.info.run_id,
+                                            MLFLOW_PROJECT_ENV, "conda")
             command_separator = " && "
-            conda_env_name = _get_or_create_conda_env(project.conda_env_path)
+            conda_env_name = _get_or_create_conda_env(fetched_project.project.conda_env_path)
             command_args += _get_conda_command(conda_env_name)
         # In synchronous mode, run the entry point command in a blocking fashion, sending status
         # updates to the tracking server when finished. Note that the run state may not be
         # persisted to the tracking server if interrupted
         if synchronous:
-            command_args += _get_entry_point_command(project, entry_point, parameters, storage_dir)
+            command_args += _get_entry_point_command(fetched_project.project,
+                                                     entry_point, parameters, storage_dir)
             command_str = command_separator.join(command_args)
-            return _run_entry_point(command_str, work_dir, experiment_id,
-                                    run_id=active_run.info.run_id)
+            return _run_entry_point(command_str, fetched_project.work_dir, experiment_id,
+                                    run_id=fetched_project.run.info.run_id)
         # Otherwise, invoke `mlflow run` in a subprocess
         return _invoke_mlflow_run_subprocess(
-            work_dir=work_dir, entry_point=entry_point, parameters=parameters,
+            work_dir=fetched_project.work_dir, entry_point=entry_point, parameters=parameters,
             experiment_id=experiment_id,
-            use_conda=use_conda, storage_dir=storage_dir, run_id=active_run.info.run_id)
+            use_conda=use_conda, storage_dir=storage_dir, run_id=fetched_project.run.info.run_id)
     elif backend_name == "kubernetes":
         from mlflow.projects import kubernetes as kb
-        tracking.MlflowClient().set_tag(active_run.info.run_id, MLFLOW_PROJECT_ENV, "docker")
-        tracking.MlflowClient().set_tag(active_run.info.run_id, MLFLOW_PROJECT_BACKEND,
+        tracking.MlflowClient().set_tag(fetched_project.run.info.run_id,
+                                        MLFLOW_PROJECT_ENV, "docker")
+        tracking.MlflowClient().set_tag(fetched_project.run.info.run_id, MLFLOW_PROJECT_BACKEND,
                                         "kubernetes")
-        _validate_docker_env(project)
+        _validate_docker_env(fetched_project.project)
         _validate_docker_installation()
         kube_config = _parse_kubernetes_config(backend_config)
-        image = _build_docker_image(work_dir=work_dir,
+        image = _build_docker_image(work_dir=fetched_project.work_dir,
                                     repository_uri=kube_config["repository-uri"],
-                                    base_image=project.docker_env.get('image'),
-                                    run_id=active_run.info.run_id)
+                                    base_image=fetched_project.project.docker_env.get('image'),
+                                    run_id=fetched_project.run.info.run_id)
         image_digest = kb.push_image_to_registry(image.tags[0])
         submitted_run = kb.run_kubernetes_job(
-            project.name,
-            active_run,
+            fetched_project.project.name,
+            fetched_project.run,
             image.tags[0],
             image_digest,
-            _get_entry_point_command(project, entry_point, parameters, storage_dir),
-            utils.generate_env_vars_to_attach_to_run(active_run),
+            _get_entry_point_command(fetched_project.project, entry_point, parameters, storage_dir),
+            utils.generate_env_vars_to_attach_to_run(fetched_project.run),
             kube_config.get('kube-context', None),
             kube_config['kube-job-template']
         )
@@ -313,87 +283,6 @@ def _wait_for(submitted_run_obj):
         submitted_run_obj.cancel()
         _maybe_set_run_terminated(active_run, "FAILED")
         raise
-
-
-def _fetch_project(uri, force_tempdir, version=None):
-    """
-    Fetch a project into a local directory, returning the path to the local project directory.
-    :param force_tempdir: If True, will fetch the project into a temporary directory. Otherwise,
-                          will fetch ZIP or Git projects into a temporary directory but simply
-                          return the path of local projects (i.e. perform a no-op for local
-                          projects).
-    """
-    parsed_uri, subdirectory = _parse_subdirectory(uri)
-    use_temp_dst_dir = force_tempdir or _is_zip_uri(parsed_uri) or not _is_local_uri(parsed_uri)
-    dst_dir = tempfile.mkdtemp() if use_temp_dst_dir else parsed_uri
-    if use_temp_dst_dir:
-        _logger.info("=== Fetching project from %s into %s ===", uri, dst_dir)
-    if _is_zip_uri(parsed_uri):
-        if _is_file_uri(parsed_uri):
-            parsed_file_uri = urllib.parse.urlparse(urllib.parse.unquote(parsed_uri))
-            parsed_uri = os.path.join(parsed_file_uri.netloc, parsed_file_uri.path)
-        _unzip_repo(zip_file=(
-            parsed_uri if _is_local_uri(parsed_uri) else _fetch_zip_repo(parsed_uri)),
-            dst_dir=dst_dir)
-    elif _is_local_uri(uri):
-        if version is not None:
-            raise ExecutionException("Setting a version is only supported for Git project URIs")
-        if use_temp_dst_dir:
-            dir_util.copy_tree(src=parsed_uri, dst=dst_dir)
-    else:
-        assert _GIT_URI_REGEX.match(parsed_uri), "Non-local URI %s should be a Git URI" % parsed_uri
-        _fetch_git_repo(parsed_uri, version, dst_dir)
-    res = os.path.abspath(os.path.join(dst_dir, subdirectory))
-    if not os.path.exists(res):
-        raise ExecutionException("Could not find subdirectory %s of %s" % (subdirectory, dst_dir))
-    return res
-
-
-def _unzip_repo(zip_file, dst_dir):
-    import zipfile
-    with zipfile.ZipFile(zip_file) as zip_in:
-        zip_in.extractall(dst_dir)
-
-
-def _fetch_zip_repo(uri):
-    import requests
-    from io import BytesIO
-    # TODO (dbczumar): Replace HTTP resolution via ``requests.get`` with an invocation of
-    # ```mlflow.data.download_uri()`` when the API supports the same set of available stores as
-    # the artifact repository (Azure, FTP, etc). See the following issue:
-    # https://github.com/mlflow/mlflow/issues/763.
-    response = requests.get(uri)
-    try:
-        response.raise_for_status()
-    except requests.HTTPError as error:
-        raise ExecutionException("Unable to retrieve ZIP file. Reason: %s" % str(error))
-    return BytesIO(response.content)
-
-
-def _fetch_git_repo(uri, version, dst_dir):
-    """
-    Clone the git repo at ``uri`` into ``dst_dir``, checking out commit ``version`` (or defaulting
-    to the head commit of the repository's master branch if version is unspecified).
-    Assumes authentication parameters are specified by the environment, e.g. by a Git credential
-    helper.
-    """
-    # We defer importing git until the last moment, because the import requires that the git
-    # executable is availble on the PATH, so we only want to fail if we actually need it.
-    import git
-    repo = git.Repo.init(dst_dir)
-    origin = repo.create_remote("origin", uri)
-    origin.fetch()
-    if version is not None:
-        try:
-            repo.git.checkout(version)
-        except git.exc.GitCommandError as e:
-            raise ExecutionException("Unable to checkout version '%s' of git repo %s"
-                                     "- please ensure that the version exists in the repo. "
-                                     "Error: %s" % (version, uri, e))
-    else:
-        repo.create_head("master", origin.refs.master)
-        repo.heads.master.checkout()
-    repo.submodule_update(init=True, recursive=True)
 
 
 def _get_conda_env_name(conda_env_path, env_id=None):
@@ -546,38 +435,6 @@ def _run_mlflow_run_cmd(mlflow_run_arr, env_map):
     else:
         return subprocess.Popen(
             mlflow_run_arr, env=final_env, universal_newlines=True, preexec_fn=os.setsid)
-
-
-def _create_run(uri, experiment_id, work_dir, entry_point):
-    """
-    Create a ``Run`` against the current MLflow tracking server, logging metadata (e.g. the URI,
-    entry point, and parameters of the project) about the run. Return an ``ActiveRun`` that can be
-    used to report additional data about the run (metrics/params) to the tracking server.
-    """
-    if _is_local_uri(uri):
-        source_name = tracking._tracking_service.utils._get_git_url_if_present(_expand_uri(uri))
-    else:
-        source_name = _expand_uri(uri)
-    source_version = _get_git_commit(work_dir)
-    existing_run = fluent.active_run()
-    if existing_run:
-        parent_run_id = existing_run.info.run_id
-    else:
-        parent_run_id = None
-
-    tags = {
-        MLFLOW_USER: _get_user(),
-        MLFLOW_SOURCE_NAME: source_name,
-        MLFLOW_SOURCE_TYPE: SourceType.to_string(SourceType.PROJECT),
-        MLFLOW_PROJECT_ENTRY_POINT: entry_point
-    }
-    if source_version is not None:
-        tags[MLFLOW_GIT_COMMIT] = source_version
-    if parent_run_id is not None:
-        tags[MLFLOW_PARENT_RUN_ID] = parent_run_id
-
-    active_run = tracking.MlflowClient().create_run(experiment_id=experiment_id, tags=tags)
-    return active_run
 
 
 def _get_run_env_vars(run_id, experiment_id):
